multi-core bug fix for assign_fixed_points
